+++ lib/octopress/isolate.rb
@@ -2,15 +2,15 @@ require 'find'
 
 module Octopress
   class Isolate
+    attr_accessor :site
+
     def initialize(options)
       @options = options
-      Jekyll.logger.log_level = :error
-      @site = Jekyll::Site.new(Jekyll.configuration(@options))
-      Jekyll.logger.log_level = :info      
+      @site = Octopress.site(@options)
     end
 
     def revert
-      dir = File.join(@site.source, '_posts')
+      dir = File.join(site.source, '_posts')
       exile_dir = File.join(dir, '_exile')
       if Dir.exist?(exile_dir)
         posts = find_exiled_posts
@@ -30,7 +30,7 @@ module Octopress
     def process
 
       if @options['path']
-        path = File.join(@site.source, @options['path'])
+        path = File.join(site.source, @options['path'])
         if File.exist? path
           isolate_except(path)
         else
@@ -45,21 +45,21 @@ module Octopress
 
     # remove all posts that do not match string
     def isolate_search(string)
-      posts = find_posts.select { |p| p =~ /#{string.gsub(/\s/, '-')}/i }
+      posts = Utils.find_posts.select { |p| p =~ /#{string.gsub(/\s/, '-')}/i }
       isolate_except(posts)
     end
 
     # Isolate all but the most recent post
     def isolate_latest
-      post = @site.read_posts('').sort_by(&:date).last
-      path = File.join(@site.source, post.path)
+      post = site.read_posts('').sort_by(&:date).last
+      path = File.join(site.source, post.path)
       isolate_except(path)
     end
 
     def isolate_except(posts)
       others = find_other_posts(posts)
       posts = default_array(posts)
-      exile_dir = @site.source, '_posts/_exile'
+      exile_dir = site.source, '_posts/_exile'
 
 
       if posts.size > 0
@@ -68,7 +68,7 @@ module Octopress
 
         puts "Isolated #{posts.size} post#{'s' if posts.size != 1}:"
         posts.each do |p|
-          puts "  - #{p.sub(@site.source+'/_posts/', '')}" 
+          puts "  - #{p.sub(site.source+'/_posts/', '')}" 
         end
         puts "Moved #{others.size} post#{'s' if others.size != 1} into _posts/_exile"
       else
@@ -79,23 +79,17 @@ module Octopress
     def find_other_posts(paths)
       paths = default_array(paths)
 
-      find_posts.reject do |p| 
+      Utils.find_posts.reject do |p| 
         paths.include?(p)
       end
     end
 
     def find_posts
-      dir = File.join(@site.source, '_posts')
-      Find.find(dir).to_a.reject do |f| 
-        File.directory?(f) || f =~ /_exile\//
-      end
+      Utils.find_posts.reject { |f| f =~ /_exile\// }
     end
 
     def find_exiled_posts
-      dir = File.join(@site.source, '_posts/_exile')
-      Find.find(dir).to_a.reject do |f| 
-        File.directory?(f)
-      end
+      Utils.find_posts.reject { |f| f !=~ /_exile\// }
     end
 
     def default_array(input)